wip: migrate to apollo-modules-node and backend-modules
integrate backend-modules-base
migrate /graphql to match expected layout by apollo-modules-node
integrate backend-modules-auth, -discussion and -scalars

TODOs
- [ ] get rid of lib/factories/createUser
- [ ] provide a clean path to migrate DB from crowdfunding-backend
- [ ] clean migrations
- [ ] use backend-modules-translate
- [ ] use backend-modules-format
- [ ] cleanup deps in package.json
