--- graphql/resolvers/JSON.js
@@ -1,3 +1,0 @@
-const GraphQLJSON = require('graphql-type-json')
-
-module.exports = GraphQLJSON
