--- graphql/resolvers/_RootQuerys/me.js
@@ -1,10 +1,0 @@
-module.exports = async (_, args, { user, pgdb }) => {
-  if (!user) {
-    return user
-  }
-
-  return {
-    ...user,
-    publicUser: user
-  }
-}
