+++ graphql/resolvers/_mutations/discussion/lib/voteComment.js
