+++ package.json
@@ -2,49 +2,33 @@
   "name": "republik-backend",
   "version": "0.0.1",
   "engines": {
-    "node": "8.9.0"
+    "node": "8.9.1",
+    "yarn": "1.3.2"
   },
   "private": true,
+  "license": "AGPL-3.0",
   "dependencies": {
-    "apollo-engine": "^0.5.2",
-    "apollo-server-express": "^1.2.0",
+    "@orbiting/backend-modules-auth": "^0.4.1",
+    "@orbiting/backend-modules-base": "^0.0.6",
+    "@orbiting/backend-modules-documents": "^0.0.4",
+    "@orbiting/backend-modules-mail": "^0.0.0",
+    "@orbiting/backend-modules-scalars": "^0.0.3",
+    "@orbiting/backend-modules-scripts": "^1.0.0",
+    "@orbiting/backend-modules-translate": "^0.0.0",
     "aws-sign": "^0.1.2",
-    "body-parser": "^1.18.2",
-    "connect-pg-simple": "^4.2.1",
-    "cookie-parser": "^1.4.3",
-    "cors": "^2.8.4",
     "d3-array": "^1.2.1",
     "d3-time-format": "^2.1.0",
     "db-migrate": "^0.10.0-beta.24",
     "db-migrate-pg": "^0.2.5",
     "dotenv": "^4.0.0",
-    "email-validator": "^1.1.1",
-    "express": "^4.16.2",
-    "express-session": "^1.15.6",
-    "geoip-database": "^1.0.9",
+    "dynamic-dedupe": "^0.3.0",
     "graphql": "^0.11.7",
     "graphql-fields": "^1.0.2",
-    "graphql-redis-subscriptions": "^1.3.1",
-    "graphql-subscriptions": "^0.5.4",
-    "graphql-tools": "^2.7.2",
-    "graphql-type-json": "^0.1.4",
+    "graphql-tools": "https://github.com/orbiting/graphql-tools.git#upstream",
     "isomorphic-unfetch": "^2.0.0",
     "json-stable-stringify": "^1.0.1",
-    "kraut": "^0.1.1",
     "lodash": "^4.17.4",
-    "maxmind": "^2.3.0",
-    "passport": "^0.4.0",
-    "pogi": "^2.4.0",
-    "query-string": "^5.0.1",
-    "redis": "^2.8.0",
-    "sharp": "^0.18.4",
-    "subscriptions-transport-ws": "^0.9.1",
-    "superb": "^2.0.0",
-    "superheroes": "^1.0.0",
-    "supervillains": "^1.0.0",
-    "tslib": "^1.8.0",
-    "unified": "^6.1.5",
-    "unist-util-visit": "^1.1.3"
+    "sharp": "^0.18.4"
   },
   "scripts": {
     "dev": "NODE_ENV=development nodemon -w graphql -w src -w lib -w .env -w package.json -w server.js -w index.js index.js",
@@ -57,7 +41,9 @@
     "db:seed": "cat seeds/seeds.json | node seeds/seed.js --truncate",
     "db:reset:noseeds": "db-migrate reset && db-migrate up",
     "db:reset": "npm run db:reset:noseeds && npm run db:seed",
-    "translations": "gsheets --key=1NE1igcbxtTrVd5P72Ubqmgv4yKYh0LRGa0EuYSEQKxE --title=live --pretty --out lib/translations.json"
+    "translations": "gsheets --key=1NE1igcbxtTrVd5P72Ubqmgv4yKYh0LRGa0EuYSEQKxE --title=live --pretty --out lib/translations.json",
+    "link:backend-modules": "yarn link apollo-modules-node && yarn link @orbiting/backend-modules-auth && yarn link @orbiting/backend-modules-base && yarn link @orbiting/backend-modules-documents && yarn link @orbiting/backend-modules-mail && yarn link @orbiting/backend-modules-translate && yarn link @orbiting/backend-modules-scalars && yarn link @orbiting/backend-modules-scripts",
+    "unlink:backend-modules": "yarn unlink apollo-modules-node && yarn unlink @orbiting/backend-modules-auth && yarn unlink @orbiting/backend-modules-base && yarn unlink @orbiting/backend-modules-documents && yarn unlink @orbiting/backend-modules-mail && yarn unlink @orbiting/backend-modules-translate && yarn unlink @orbiting/backend-modules-scalars && yarn unlink @orbiting/backend-modules-scripts"
   },
   "lint-staged": {
     "*.js": [
