+++ server.js
@@ -1,93 +1,84 @@
-const PgDb = require('./lib/pgdb')
-const express = require('express')
-const cors = require('cors')
-const { createServer } = require('http')
-const { Engine } = require('apollo-engine')
+const { server } = require('@orbiting/backend-modules-base')
 
-const DEV = process.env.NODE_ENV && process.env.NODE_ENV !== 'production'
-if (DEV) {
-  require('dotenv').config()
+// schema stitching has it's pitfalls (in the current version), when one
+// schema extends another (local extends User of backend-modules-auth).
+//   - we need to merge types manually here (with mergeObjectTypes)
+//   - we need to provide the local resolvers to mergeSchemas resolvers,
+//     otherwise the local resolvers don't get called, even tough mergeObjectTypes
+//     seems to properly merge them :(
+//
+// [RFC] Next steps for schema stitching
+// https://github.com/apollographql/graphql-tools/issues/495
+//
+// mergeSchemas breaks the recursive nature of resolvers
+// https://github.com/apollographql/graphql-tools/issues/443
+//
+// mergeSchema - onTypeConflict shouldn't be called with built-in types
+// https://github.com/apollographql/graphql-tools/issues/487
+
+/*
+const cloneDeep = require('lodash.clonedeep')
+const { GraphQLObjectType } = require('graphql/type/definition')
+const { mergeSchemas } = require('graphql-tools')
+
+function mergeObjectTypes (leftType, rightType) {
+  if (!rightType) {
+    return leftType
+  }
+  if (leftType.constructor.name !== rightType.constructor.name) {
+    throw new TypeError(`Cannot merge with different base type. this: ${leftType.constructor.name}, other: ${rightType.constructor.name}.`)
+  }
+  const mergedType = cloneDeep(leftType)
+  mergedType.getFields() // Populate _fields
+  for (const [key, value] of Object.entries(rightType.getFields())) {
+    mergedType._fields[key] = value
+  }
+  if (leftType instanceof GraphQLObjectType) {
+    mergedType._interfaces = Array.from(new Set(leftType.getInterfaces().concat(rightType.getInterfaces())))
+  }
+  //console.log('mergedType:', mergedType.getFields())
+  return mergedType
 }
 
-process.env.PORT = process.env.PORT || 3020
+const mergeTypes = [ 'User' ]
+*/
 
-const {
-  PORT,
-  CORS_WHITELIST_URL,
-  SESSION_SECRET,
-  COOKIE_DOMAIN,
-  ENGINE_API_KEY
-} = process.env
-
-const auth = require('./src/auth')
-const graphql = require('./graphql')
-
-let pgdb
-let server
-let httpServer
-let subscriptionServer
 module.exports.run = () => {
-  // init apollo engine
-  const engine = ENGINE_API_KEY
-    ? new Engine({
-      engineConfig: {
-        apiKey: ENGINE_API_KEY,
-        logging: {
-          level: 'INFO'   // Engine Proxy logging level. DEBUG, INFO, WARN or ERROR
-        }
-      },
-      graphqlPort: PORT
-    })
-    : null
-  if (engine) {
-    engine.start()
-  }
+  // const { makeExecutableSchema, mergeSchemas } = require('graphql-tools')
+  const { makeExecutableSchema } = require('graphql-tools')
+  const t = require('./lib/t')
 
-  return PgDb.connect().then((_pgdb) => {
-    pgdb = _pgdb
-    server = express()
-    httpServer = createServer(server)
+  const localModule = require('./graphql')
 
-    // apollo engine middleware
-    if (engine) {
-      server.use(engine.expressMiddleware())
-    }
-
-    // Once DB is available, setup sessions and routes for authentication
-    auth.configure({
-      server: server,
-      secret: SESSION_SECRET,
-      domain: COOKIE_DOMAIN || undefined,
-      dev: DEV,
-      pgdb: pgdb
-    })
-
-    if (CORS_WHITELIST_URL) {
-      const corsOptions = {
-        origin: CORS_WHITELIST_URL.split(','),
-        credentials: true,
-        optionsSuccessStatus: 200 // some legacy browsers (IE11, various SmartTVs) choke on 204
-      }
-      server.use('*', cors(corsOptions))
-    }
-
-    subscriptionServer = graphql(server, pgdb, httpServer)
-
-    // start the server
-    httpServer.listen(PORT, () => {
-      console.info('server is running on http://localhost:' + PORT)
-    })
-
-    return { pgdb }
+  // graphql schema
+  /*
+  const executableSchema = mergeSchemas({
+    schemas: [
+      makeExecutableSchema(require('@orbiting/backend-modules-auth').graphql),
+      makeExecutableSchema(require('@orbiting/backend-modules-documents').graphql),
+      makeExecutableSchema(localModule),
+    ],
+    //onTypeConflict: (leftType, rightType) => {
+    //  if (mergeTypes.indexOf(leftType.toString()) > -1 && leftType instanceof GraphQLObjectType) {
+    //    return mergeObjectTypes(leftType, rightType)
+    //  }
+    //  return leftType
+    //},
+    //resolvers: mergeInfo => localModule.typeResolvers
   })
+  */
+  const { merge } = require('apollo-modules-node')
+
+  const { graphql: documents } = require('@orbiting/backend-modules-documents')
+
+  const executableSchema = makeExecutableSchema(merge(localModule, [documents]))
+
+  // middlewares
+  const middlewares = []
+
+  return server.run(executableSchema, middlewares, t)
 }
 
 module.exports.close = () => {
-  const { pubsub } = require('./lib/RedisPubSub')
-  pubsub.getSubscriber().quit()
-  pubsub.getPublisher().quit()
-  subscriptionServer.close()
-  httpServer.close()
-  pgdb.close()
-  require('./lib/redis').quit()
+  server.close()
 }
